Latest

+++ what2craft/src/main/java/database/ItemRepository.java
@@ -1,10 +1,11 @@
 package database;
 
 import org.springframework.data.jpa.repository.JpaRepository;
+import org.springframework.stereotype.Repository;
 
 import logic.Item;
 
-public interface ItemRepository extends JpaRepository<Item, Long>{
+public interface ItemRepository extends JpaRepository<Item, Integer>{
